euler Square root convergents

+++ src/main/kotlin/euler/Task33DigitCancellingFractions.kt
@@ -23,7 +23,7 @@ object Task33DigitCancellingFractions : EulerTask("Digit cancelling fractions") 
             }
         }
 
-        return product.simplified.denominator
+        return product.simplified.denominator.toLong()
     }
 
     private fun anomalousCancellation(fraction: Fraction): Boolean {
